% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{print.OpenDALMetadata}
\alias{print.OpenDALMetadata}
\title{S3 print method for OpenDALMetadata objects}
\usage{
\method{print}{OpenDALMetadata}(x, ...)
}
\arguments{
\item{x}{An OpenDALMetadata object.}

\item{...}{Further arguments passed to or from other methods (unused).}
}
\description{
S3 print method for OpenDALMetadata objects
}
\keyword{internal}
