% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{print.OpenDALOperator}
\alias{print.OpenDALOperator}
\title{S3 print method for OpenDALOperator objects}
\usage{
\method{print}{OpenDALOperator}(x, ...)
}
\arguments{
\item{x}{An OpenDALOperator object.}

\item{...}{Further arguments passed to or from other methods (unused).}
}
\description{
S3 print method for OpenDALOperator objects
}
\keyword{internal}
